// src/components/navigation/tab-bar/tab-bar-item-type.tsx
import type { ReactNode } from 'react';

export type TabBarItem = {
  /**
   * The text label for the tab.
   */
  label: string;
  /**
   * Unique identifier for the tab.
   */
  value: string;
  /**
   * Icon to display alongside the label.
   */
  icon?: ReactNode;
  /**
   * Optional URL if the tab should act as a link.
   */
  href?: string;
};